Fix submit-form 422: completionTimeInSeconds must be an integer
The backend deserializes formDetails.completionTimeInSeconds as i32, but the
SDKs typed it Double. kotlinx.serialization emits whole doubles as "1.0",
so every Android/KMP/Compose submit with a completion time got a 422. Swift's
JSONEncoder happens to emit "1" for whole doubles, masking the bug on iOS —
but a fractional value from the web form would 422 there too.

Now Int across core/kmp-sdk/ios-native wire types and builder options, with
rounding at both WebView-bridge parse sites so fractional web-reported
seconds can never reach the wire. dist/ rebuilt.

+++ core/src/commonMain/kotlin/com/encatch/core/FormSubmitBuilder.kt
@@ -30,7 +30,7 @@ data class BuildSubmitRequestOptions(
     val formConfigurationId: String,
     val triggerType: String = "manual",
     val responseLanguageCode: String? = null,
-    val completionTimeInSeconds: Double? = null,
+    val completionTimeInSeconds: Int? = null,
     val isPartialSubmit: Boolean? = null,
     val feedbackIdentifier: String? = null,
     /** Arbitrary caller-provided metadata attached to this submission. */
